Windows colors: Fixed remaining problems and refactored
This is unfortunately large commit (missing hg) and the diff is made even
bigger by eol changes. The highlights of these highlight changes are:

1) Fixed problem getting original console colors. GetConsoleScreenBufferInfo
got handle in wrong format and thus it failed and didn't set correct colors.
Now the method is called correctly and a check for returned status was added
as a safeguard.

2) Replaced HighlighterFor factory with Highlighter class that uses OS specific
highlighters internally. The new class handles highlighting both stdout and
stderr and it is possible to create it only once in monitor. This class made
special NoHighlighting class unnecessary.

3) Made OS specific highlighters dummier. Nowadays they only set requested
color and can reset it later. Mapping messages to colors is done in by the
new Highlighter class.

I think the code is pretty OK now but review is needed. I might also want to
move UnixHighligher and DosHighlighter into same module and think could they
be named better.

Update issue 753
Status: Done
Cc: pekka.klarck,janne.t.harkonen

This issue should now be done otherwise but testing on Linux and code review
are needed. You still need to use "--MonitorColors force" to enable colors on
Windows, but that limitation will be lifted soon when issue 777 is done.