Warnings when registering signal handlers fail.
Update issue 816
Status: Done
Owner: pekka.klarck

Nowadays there is a warning if registering a signal handler fails. Depending on the signal the error is either

    Registering signal INT failed. Stopping execution gracefully with this signal or with Ctrl-C is not possible. Original error was: <message>

or

    Registering signal TERM failed. Stopping execution gracefully with this signal is not possible. Original error was: <message>

I decided not to emit a warning if Robot doesn't run on the main thread. There is no way to make signals work in that case so there should have been a way to disable this warning somehow.