API and UG docs for robot.api.logger
Update issue 339
Status: Review
Cc: janne.t.harkonen
First of all, we have agreed to keep the new API in `robot.api.logger`. We'll also use `robot.api` for other public APIs in the future.

API docs written and a short section about the new logging API added to the User Guide. Janne, could you review the docs?

The API docs still needs to be linked from the User Guide but that's covered by issue 482 (and a FIXME I added to the UG).