Refactored import logic and at the same time fixed problem with importing two libraries with same name using physical path
Update issue 822
Status: Done
The problem is now fixed (or at least the test I created earlier passes).

The code changes were quite a bit larger than absolutely needed, but I wanted
to cleanup the import logic at the same time. After the refactoring libs
specified with a path are imported using the same code that is used for
importing variable files which already had a fix for importing module with
a same name as an earlier imported different module.