Fixed showing combined pattern and links for tags.
The former was broken because attribute was renamed.

The latter was broken with totals because jQuery templates apparently
evaluate both sides of {{if links && links.length}} and this obviously
fails when links is undefined. {{if links}}{{if links.length}} works.

It's highly annoying this stuff cannot be tested without browser tests.