Inplace fix for jquery tmpl bug with escaping &.
The fix was done based on the comments on the issue tracker [1] and basically

   .split("&").join("&amp;")

was added before

   .split("<").join("&lt;").split(">").join("&gt;"). ...

We don't show non-ASCII chars like &auml; (we use UTF-8 instead) so
the concerns at [1] aren't valid for us.

Now we just need to remember to apply the same fix if/when this file
is updated (unless the actual bug is already fixed).

[1] https://github.com/jquery/jquery-tmpl/issues/79