Explain how custom embedded arg regexps automatically match variables.
Update issue 854
Status: Review
Now also handling variables automatically is explained in the User
Guide.

This issue ought to be ready except for review. Because we have pretty
comprehensive tests I'm not too worried about the code, but
proofreading the documentation would be highly appreciated. It's probably
best to read the whole embedded arguments section that is linked below:

http://robotframework.googlecode.com/hg/doc/userguide/RobotFrameworkUserGuide.html#embedding-arguments-into-keyword-name