Refactored getting resources under robot.webcontent
Update issue 916
Status: Started
Cc: pekka.klarck
I did some refactoring to the code. It works with Python and normal Jython
but I don't remember how to build robotframework.jar to test that (and it's
too late that I don't bother to find out). The code thus needs more testing
and also review.