Better test data for manually testing dynamic __intro__ and __init__.
Update issue 186
I did some manual testing and noticed __init__ arguments do not work.
I looked at the code and it seems that this functionality is still missing.
Most likely this needs to be handled with argsgetter similar to docgetter
used for getting the __init__ doc.

This commit contains changes to libdoc's and remote server's examples that
can be used for manually testing this. Creating automated end-to-end or
integration tests would be great too.