Fix problems caused by listener methods logging by writing messages outside keywords into syslog instead of normal log.
Update issue 548
Status: Started
Owner: pekka.klarck
The problems with listener methods logging causing the output to be corrupt
ought to be fixed now. The fix was changing logging logic so that messages
logged when keywords are not executed are written into the syslog instead
of the normal log. Possible warnings are logged into the execution errors
section of the log file, though.

The listener methods that can log something so that it appears in the normal
log file are start_keyword, end_keyword, log_message and message (when it is
called while keywords are executed). As already mentioned, messages by other
listener methods are written to syslog.

This commit contains implementation that ought to be pretty much ready. I have
also started creating tests for this but don't have time to finalize them
right now. Finally, also the User Guide should be updated to explain how to
log from listeners and what limitations it has.