Implementation and tests for automatic ${KEYWORD STATUS} and ${KEYWORD MESSAGE} variables.
Update issue 1040
Status: Started
Owner: pekka.klarck
I implemented ${KEYWORD STATUS} and also ${KEYWORD MESSAGE} variable. They
are available in the keyword teardown both when used directly and in user
keywords used in teardowns.

The implementation was pretty simple and also test coverage ought to be good
enough. Review is still needed and this also must be documented in UG.