Added tests for --TagStatInclude/Exclude with rebot
Update issue 1072
Added tests for this also when using rebot. As I suspected, that implementation was correct and no further fixes needed.

It is highly annoying that we currently have to statistics collection implementations in our code. Unfortunatley removing the old one that is used when running tests needs to wait until test execution refactoring in RF 2.8.