Support options that can be given multiple times on cli also as single strings programmatically.
Update issue 1103
Status: Started
Owner: pekka.klarck
Labels: Target-2.7.2
Decided to include this into 2.7.2 because this is so easy to implement.

Now using options that can be given multiple times on cli can be given as
a single string when using robot.run or robot.rebot. Adding similar support
for robot.running.model.TestSuite is trivial but I need to figure out where
to add a test for it. That can wait until tomorrow.