Changed unknown test library version from '<unknown>' to ''.
The change was done mainly to avoid showing '<unknown>' in docs generated by libdoc. It already excluded version altogether when there was none (which is the case with resource files).

When test library import is syslogged at run time, unknown version is still logged as '<unknown>'. This makes that message easier to understand (and avoided changing tests related to it).

Added also explicit tests for handling library versions with libdoc.