Result building: Omit keywords if building only report.
Update issue 1261
Status: Started
Changed the code so that if only report is generated, keywords in output.xml
are nor processed at all. I want to still refactor the code a little and
thus I don't close this issue yet.

The performance impact of the change is pretty huge. In the case mentioned
in the original description, total time is now 1 minute and maximum memory
63 MB.