Fixed user keywords returning correct value also when they contain continuable failures
Update issue 1266
Status: Done
Luckily fixing this wasn't too complicated. Just needed to make it possible that exceptions used for reporting failures can also contain return values. Then if an error is "continuable", we can get the return value directly from the error.

I'm not entirely happy about the code. Didn't want to spent too much refactoring it, though, because this is likely to be totally rewritten in RF 2.8 or later. The added tests ought to make it sure that this functionality will not break during refactoring.