Refactored handling test execution errors.
The original motivation was to clean-up the code so that implementing issue 1307 would be easier. The code turned out to be a can of worms but opening it was worth the effort.

As part of the refactoring I removed some unnecessary instance variables containing references to result and context objects. Earlier these objects were never garbage collected which resulted to a memory leak reported already earlier as issue 1202.

Update issue 1202
Status: Review
Labels: Target-2.7.6
This reather big refactoring/clean-up commit removes the unnecessary references to result and context objects that caused the memory leak.

I leave the issue still open because I want to go through the code with fresh eyes.