Add the actuall error message to 'parent suite failed' errors.
Update issue 1307
Now also suite teardowns are handled. This needed a bit more work
because when suite teardowns fail, all test related information has
already been written to output.xml. The message that suite teardown has
failed is added only when output is processed. There was no easy way to
get the correct errros at this point, so we needed to add it to
output.xml.

Now when a teardown fails, the error message is written into <status>
element similarly as with tests and suites. Other keywords don't need
this information so it was not added to them. Test teardowns don't
really need it either, but there was no simple/fast way to check is a
keyword suite or test teardown.