Telnet.Read Until: Verify that the output contains expected before decoding.
Update issue 1328
After we had started to use UTF-8 and ignore as the default encoding parameters, this Python bug hit us:
http://bugs.python.org/issue8271

In short that bug meant that decoding '\xf21234' resulted with '4' by default when you would expect '1234'. This caused problems when Read Until was comparing that the expected value had actually was found from the returned output.

Luckily avoiding this bug is pretty simple as we only need to verify that output contains the extected string before decoding it. As an added bonus, the code also got more clean.