Fixed using ... as value in settings table.
Update issue 1379
Status: Done
Moved the fix from VariablePopulator to _PropertyPopulator to get it also to
SettingPopulator. Also used a different, and hopefully a bit more clear,
solution to fix the problem. Using a flag is always a bit ugly, but in this
case it helps understanding the intention of the code.

Also added tests for handling ... in settings table.