Parsing: Added atests for handling escaping using `\` with different formats.
Update issue 1495
Noticed that escaping is done differently with different reST parsers and
added tests to verify it.

The old parser requires double escaping because `\` is an escape character
both in Robot and in reST. For example, Windows paths need to be given like
`C:\\\\temp` and newline is `\\n`.

With the new parser normal escaping is enough. Windows paths look like
`C:\\Temp` and newline is `\n`.

These differences need to be documented in the User Guide. I'm currently
enhancing the docs related to reST parsing in general and I'll take care of
that too.