Added support for timeout in Wait For Process
update issue 1481
To allow generic usage of the keyword we decided that by default timeout
only exits the keyword and does nothing to the process. This allows to explicitly run keywords
before shutting the process down after timeout. Also a parameter was implemented
to allow terminating or killing the process directly on timeout.

Currently there is a 5 second timeout for termination, after which force kill is used if the process does not respond. The
arbitrary 5s might be too much (e.g. we know that the process should shut down immediately) or too little (e.g. web server)
in some circumstances so maybe this should be configurable too or then removed alltogether.