Fixed double escaping of --tagstatling url
Update issue 1494
Status: Done
Owner: pekka.klarck
Escaping was done right on Robot side but then values were escaped also in log and report templates. Changing templates to show these values as-is was enough to fix the problem.

We don't have automated tests for log/report UI, but I modified the data we use for manual testing them so that --tagstatlink is used with &.