String.Split String From Right: Fixed handling longer separators
Update issue 1553
Status: Done
Owner: pekka.klarck
Labels: bwic
Thanks for the report. Easiest solution was using rsplit as you suggested.

This was clearly not too well tested because such a bug was lurking
here. Apparently this keyword isn't used too often with longer
separators either as nobody had spotted the bug earlier. Or
alternatively they have decided to reverse the separator. In that case
this fix is backwards incompatible for them. I added `bwic` label so
that we remember to mention this in 2.8.2 release notes.