Tag patterns: Support for OR in Python side.
Update issue 1558
Status: Started
Owner: pekka.klarck
Implemented the support on Python side with help of unit tests. This seems to work already e.g. with --include, but that needs to be separately tested on higher level. Other tasks include implementing this also on JavaScript and documentation in UG and possibly also in --help.